+++ samples/complexLogic.ts
@@ -45,8 +45,9 @@ function newAction(element: HTMLElement | null): Partial<ActionType> {
 export const complexLogic = kea<
     complexLogicType<ActionForm, ActionType, AntdFieldData, DashboardItemType, FormInstance>
 >({
+    path: ['complexLogic'],
     props: {} as AntdFieldData,
-    
+
     actions: {
         setForm: (form: FormInstance) => ({ form }),
         selectAction: (id: string | null) => ({ id: id || null }),
@@ -155,8 +156,8 @@ export const complexLogic = kea<
                 const a = {} as DashboardItemType
                 const result = { ...a, bla: true }
                 return result
-            }
-        ]
+            },
+        ],
     },
 
     listeners: ({ actions, values }) => ({