+++ samples/windowValuesLogic.ts
@@ -1,6 +1,7 @@
 import { kea } from 'kea'
 import { windowValuesLogicType } from './windowValuesLogicType'
 export const windowValuesLogic = kea<windowValuesLogicType>({
+    path: ['windowValuesLogic'],
     windowValues: () => ({
         windowHeight: (window) => window.innerHeight,
         windowWidth: (window) => Math.min(window.innerWidth, window.document.body.clientWidth),